requirements.txt
python-dotenv
google-generativeai
crewai
crewai-tools